Only restore dialog focus if focus is in the dialog
Context: https://github.com/whatwg/html/issues/8904

This patch prevents the behavior where closing a dialog focuses the
previously focused element from before the dialog was opened only if
focus is in the dialog when the dialog closes.

Without this, focus can unexpectedly shift away from an element which is
not going away, like a text input that the user is currently typing into
for example.

This matches the behavior we recently implemented for popovers.

Change-Id: Id95cacf36b2eb24d3514f59bc8a1f718d643860e