Make element references via idrefs functional when disconnected
Per the spec [1], computing the attr-associated element does not
depend on connectedness. Step 4 (which runs whenever there is *not*
an explicitly set attr-element) says:

-----
...return the first element candidate, in tree order, that meets the following criteria:

 - candidate's root is the same as element's root,
 - candidate's ID is contentAttributeValue, and
 - candidate implements T.

If no such element exists, then return null.
-----

The issue prior to this CL is that `getElementById()` uses the
TreeOrderedMap, and elements are removed from the map when they are
removed from the document. A partial fix for non-idref (IDL) element
references was made with [2], and this is the second half.

[1] https://html.spec.whatwg.org/#attr-associated-element
[2] https://chromium-review.googlesource.com/c/chromium/src/+/4371122

Fixed:1427374
Change-Id: I2a089a5fa1a925d40761ae2275ed34f855791513