Pass bitmap images as mailboxes between processes
This CL adds mojo structs to pass the AcceleratedStaticImageBitmaps via
mailboxes instead of serializing it as array of pixels, which first requires
a GPU readback and inevitable upload to the GPU process on the receiver side.

A callback, wrapped mojo::PendingRemote, is passed along with the serialized image, so that the sending process would keep the original image alive until the transfer completes.

Small changes to the StaticBitmapImage interface and AcceleratedStaticBitmapImage class are done to expose information required for transfer. A new factory method is also added to reconstruct the image on the receive side.

A new WPT test exercising image sharing across domains is added to "oopr-canvas2d" and "gpu" test suites.

Bug:1396243
Change-Id: If05918ee256d59542e6fc8125e25a94748bf47b4