Re-layout writing-mode roots on ICB size change, when needed.
The inline-size of an orthogonal writing-mode root may depend on the
size of the initial containing block.

This hasn't worked flawlessly for some time (ever?). We have
orthogonal_writing_mode_root_list_ in LocalFrameView, and when the
viewport size changes, we should go through that list and mark
everything for layout. However, this will only happpen when
StyleResolver::WasViewportResized() is true, which is only triggered if
the document has any element with viewport units. That's probably
fixable, so that the flag is always set when needed.

But since orthogonal_writing_mode_root_list_ is less useful anyway with
the legacy engine gone, we're going to remove it. So here's a different
solution, purely on the layout side.

Change-Id: I7754ad87f881ba6569fc27065641a9c809e8ef86